clean up some import paths, rename isImageFetching to imageFetching
