+++ packages/aws-amplify-react-native/src/InAppMessaging/components/hooks/useMessageImage/types.ts
@@ -20,6 +20,6 @@ export type ImageLoadingState = 'loading' | 'loaded' | 'failed';
 
 export type UseMessageImage = {
 	imageDimensions: ImageDimensions;
-	isImageFetching: boolean;
+	imageFetching: boolean;
 	renderImage: boolean;
 };
