+++ packages/aws-amplify-react-native/src/InAppMessaging/components/hooks/useMessageImage/useMessageImage.ts
@@ -38,7 +38,7 @@ export default function useMessageImage(image: InAppMessageImage, layout: InAppM
 	const hasSetDimensions = !isNull(imageDimensions);
 	const hasImage = !!src;
 
-	const isImageFetching = hasImage && !hasSetDimensions;
+	const imageFetching = hasImage && !hasSetDimensions;
 	const renderImage = hasImage && hasSetDimensions;
 
 	useEffect(() => {
@@ -66,5 +66,5 @@ export default function useMessageImage(image: InAppMessageImage, layout: InAppM
 		}
 	}, [hasImage, layout, src]);
 
-	return { imageDimensions, isImageFetching, renderImage };
+	return { imageDimensions, imageFetching, renderImage };
 }
