+++ packages/aws-amplify-react-native/src/InAppMessaging/components/index.ts
@@ -2,13 +2,4 @@ export { default as BannerMessage, BannerMessageProps } from './BannerMessage';
 export { CarouselMessageProps } from './CarouselMessage';
 export { default as FullScreenMessage, FullScreenMessageProps } from './FullScreenMessage';
 export { default as InAppMessageDisplay } from './InAppMessageDisplay';
-export {
-	InAppMessageComponentBaseProps,
-	InAppMessageComponentBaseStyle,
-	InAppMessageComponentButtonProps,
-	InAppMessageComponentButtonStyle,
-	InAppMessageComponentContentProps,
-	InAppMessageComponentPosition,
-	InAppMessageComponentStyle,
-	InAppMessageComponentStyles,
-} from './types';
+export { InAppMessageComponentPosition, InAppMessageComponentStyle, InAppMessageComponentStyles } from './types';
